project/requirements.txt
anyio==4.0.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
backports.functools-lru-cache==1.6.5
beautifulsoup4==4.12.2
bleach==6.1.0
Bottleneck==1.3.7
Brotli==1.0.9
cached-property==1.5.2
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.1.4
contourpy==1.0.5
cryptography==41.0.3
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.1
entrypoints==0.4
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.18.1
Flask==2.2.2
FlightRadarAPI==1.3.12
fonttools==4.25.0
fqdn==1.5.1
google-api-core==2.16.2
google-auth==2.27.0
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-language==2.13.3
google-cloud-storage==2.14.0
google-cloud-translate==3.15.3
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
graphviz==0.20.1
grpcio==1.63.0
grpcio-status==1.62.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
iniconfig==1.1.1
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.0.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==7.4.9
jupyter-console==6.6.3
jupyter_core==5.5.0
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.1
jupyterlab-widgets==3.0.9
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mistune==3.0.2
munkres==1.1.4
mysql-connector-python==8.0.32
nbclient==0.8.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numexpr==2.8.7
numpy==1.24.3
opensky-api==1.3.0
overrides==7.4.0
packaging==23.1
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pip==23.3
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
pluggy==1.0.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pyobjc-core==10.0
pyobjc-framework-Cocoa==10.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.4.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.5.0
QtPy==2.4.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.12.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.13.0
Send2Trash==1.8.2
setuptools==68.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.2
tabulate==0.8.10
terminado==0.17.1
textblob==0.15.3
threadpoolctl==2.2.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.2
traitlets==5.13.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
typing-utils==0.1.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.7
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==2.2.3
wheel==0.41.2
widgetsnbextension==4.0.9
zipp==3.17.0
